requirements.txt
aiohttp==1.0.5
aiohttp-jinja2==0.13.0
async-timeout==1.1.0
cffi==1.9.1
chardet==2.3.0
discord.py==0.16.2
Jinja2==2.9.4
MarkupSafe==0.23
multidict==2.1.4
pycparser==2.17
PyNaCl==1.0.1
pytz==2016.10
redis==2.10.5
six==1.10.0
websockets==3.2
